refactor(ActivityLog): use dedicated component for description
